Pass the context from the HTTP handler down to the database. So if a HTTP request gets cancled, then SLS will stop processing requests.
This is also a preventive measure if database transactions get dead
locked, like if a while processing data from a sql query, another is
started to get some more information and that one is blocked, which
causes the orginal query to become deadlocked. If the user cancels the
request with the passed context these database connections will be
cleaned up and other requests can come in. Without it SLS will become
deadlocked and no new requests can be made