Parameterizing the build for upstream registry
This set of changes introduces a configuration file for the Makefile as
well as an environment variable that can be used to control where upstream
containers are pulled from.

The default upstream registry is still the HPE Artifactory, but teams that prefer
to use an alternate registry can override the default with the REGISTRY_HOST environment variable

For example, if you have a private ECR, you can substitue that with a line like this:

REGISTRY_HOST=716215456011.dkr.ecr.us-east-1.amazonaws.com make

There are also some creature comforts in the Makefile to make it easier to read and work with.